requirements.txt
pillow==10.3.0
pycryptodome==3.20.0
